# Tools/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "JongSnam.Services",
    "version": "v1"
  },
  "paths": {
...
    "/Login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "Login",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "user",
            "type": "string"
          },
          {
            "in": "query",
            "name": "password",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserDto"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
    "/Logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "Logout",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
...
    "/api/Field/Search": {
      "get": {
        "tags": [
          "Field"
        ],
        "operationId": "GetFieldBySearch",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "StartPrice",
            "type": "number",
            "format": "double"
          },
          {
            "in": "query",
            "name": "ToPrice",
            "type": "number",
            "format": "double"
          },
          {
            "in": "query",
            "name": "DistrictId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "ProvinceId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "currentPage",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FieldDtoBasePagingDto"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
    "/api/Field/{id}": {
      "get": {
        "tags": [
          "Field"
        ],
        "operationId": "GetFieldById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FieldDetailDto"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
    "/api/Field/{storeId}/Fields": {
      "get": {
        "tags": [
          "Field"
        ],
        "operationId": "GetFieldByStore",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "storeId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "currentPage",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FieldDtoBasePagingDto"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
    "/api/Field": {
      "post": {
        "tags": [
          "Field"
        ],
        "operationId": "AddField",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddFieldRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Field"
        ],
        "operationId": "UpdateField",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateFieldRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
    "/api/Field/{id}/DeleteField": {
      "put": {
        "tags": [
          "Field"
        ],
        "operationId": "DeleteField",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
...
    "/api/User": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "GetUsers",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/UserDto"
              }
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "CreateUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ProblemsDetailDto": {
      "type": "object",
      "properties": {
        "detail": {
          "type": "string"
        },
        "extensions": {
          "type": "object",
          "additionalProperties": { },
          "readOnly": true
        },
        "instance": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "errorCode": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "UserDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "contactMobile": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "userType": {
          "type": "string"
        },
        "isLoggedIn": {
          "type": "boolean"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
    "StoreModel": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "createdDate": {
          "format": "date-time",
          "type": "string"
        },
        "updatedDate": {
          "format": "date-time",
          "type": "string"
        },
        "ownerId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "subDistrictId": {
          "format": "int32",
          "type": "integer"
        },
        "districtId": {
          "format": "int32",
          "type": "integer"
        },
        "provinceId": {
          "format": "int32",
          "type": "integer"
        },
        "contactMobile": {
          "type": "string"
        },
        "latitude": {
          "format": "double",
          "type": "number"
        },
        "longtitude": {
          "format": "double",
          "type": "number"
        },
        "rules": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "isOpen": {
          "type": "boolean"
        },
        "officeHours": {
          "type": "string"
        }
      }
    },
    "ImageFieldModel": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "createdDate": {
          "format": "date-time",
          "type": "string"
        },
        "updatedDate": {
          "format": "date-time",
          "type": "string"
        },
        "fieldId": {
          "format": "int32",
          "type": "integer"
        },
        "image": {
          "type": "string"
        }
      }
    },
    "FieldDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "format": "int32",
          "type": "integer"
        },
        "storeName": {
          "type": "string"
        },
        "size": {
          "type": "string"
        },
        "isOpen": {
          "type": "boolean"
        },
        "active": {
          "type": "boolean"
        },
        "storeModel": {
          "$ref": "#/definitions/StoreModel"
        },
        "imageFieldModel": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ImageFieldModel"
          }
        }
      }
    },
    "FieldDtoBasePagingDto": {
      "type": "object",
      "properties": {
        "currentPage": {
          "format": "int32",
          "type": "integer"
        },
        "totalPage": {
          "format": "int32",
          "type": "integer"
        },
        "collection": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/FieldDto"
          }
        }
      }
    },
    "DiscountModel": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "createdDate": {
          "format": "date-time",
          "type": "string"
        },
        "updatedDate": {
          "format": "date-time",
          "type": "string"
        },
        "fieldId": {
          "format": "int32",
          "type": "integer"
        },
        "startDate": {
          "format": "date-time",
          "type": "string"
        },
        "endDate": {
          "format": "date-time",
          "type": "string"
        },
        "percentage": {
          "format": "double",
          "type": "number"
        },
        "detail": {
          "type": "string"
        }
      }
    },
    "FieldDetailDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "size": {
          "type": "string"
        },
        "isOpen": {
          "type": "boolean"
        },
        "discountModel": {
          "$ref": "#/definitions/DiscountModel"
        },
        "imageFieldModel": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ImageFieldModel"
          }
        },
        "active": {
          "type": "boolean"
        }
      }
    },
    "FieldRequest": {
      "required": [
        "name",
        "price",
        "size",
        "storeId"
      ],
      "type": "object",
      "properties": {
        "storeId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 50,
          "type": "string"
        },
        "size": {
          "maxLength": 50,
          "type": "string"
        },
        "price": {
          "format": "int32",
          "type": "integer"
        },
        "isOpen": {
          "type": "boolean"
        },
        "active": {
          "type": "boolean"
        }
      }
    },
    "DiscountRequest": {
      "required": [
        "endDate",
        "percentage",
        "startDate"
      ],
      "type": "object",
      "properties": {
        "startDate": {
          "format": "date-time",
          "type": "string"
        },
        "endDate": {
          "format": "date-time",
          "type": "string"
        },
        "percentage": {
          "format": "double",
          "maximum": 100,
          "minimum": 0,
          "type": "number"
        },
        "detail": {
          "maxLength": 50,
          "type": "string"
        }
      }
    },
    "ImageFieldRequest": {
      "required": [
        "image"
      ],
      "type": "object",
      "properties": {
        "image": {
          "type": "string"
        }
      }
    },
    "AddFieldRequest": {
      "type": "object",
      "properties": {
        "fieldRequest": {
          "$ref": "#/definitions/FieldRequest"
        },
        "discountRequest": {
          "$ref": "#/definitions/DiscountRequest"
        },
        "pictureFieldRequest": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ImageFieldRequest"
          }
        }
      }
    },
    "UpdateDiscountRequest": {
      "required": [
        "endDate",
        "percentage",
        "startDate"
      ],
      "type": "object",
      "properties": {
        "startDate": {
          "format": "date-time",
          "type": "string"
        },
        "endDate": {
          "format": "date-time",
          "type": "string"
        },
        "percentage": {
          "format": "double",
          "maximum": 100,
          "minimum": 0,
          "type": "number"
        },
        "detail": {
          "maxLength": 50,
          "type": "string"
        },
        "id": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "UpdatePictureFieldRequest": {
      "required": [
        "image"
      ],
      "type": "object",
      "properties": {
        "image": {
          "type": "string"
        },
        "id": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "UpdateFieldRequest": {
      "required": [
        "name",
        "price",
        "size",
        "storeId"
      ],
      "type": "object",
      "properties": {
        "storeId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 50,
          "type": "string"
        },
        "size": {
          "maxLength": 50,
          "type": "string"
        },
        "price": {
          "format": "int32",
          "type": "integer"
        },
        "isOpen": {
          "type": "boolean"
        },
        "active": {
          "type": "boolean"
        },
        "updateDiscountRequest": {
          "$ref": "#/definitions/UpdateDiscountRequest"
        },
        "updatePictureFieldRequest": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UpdatePictureFieldRequest"
          }
        }
      }
    },
...
    "UserRequest": {
      "required": [
        "confirmPassword",
        "contactMobile",
        "email",
        "firstName",
        "lastName",
        "password",
        "userTypeId"
      ],
      "type": "object",
      "properties": {
        "email": {
          "format": "email",
          "maxLength": 50,
          "type": "string"
        },
        "password": {
          "maxLength": 20,
          "minLength": 8,
          "type": "string"
        },
        "confirmPassword": {
          "maxLength": 20,
          "minLength": 8,
          "type": "string"
        },
        "firstName": {
          "format": "string",
          "maxLength": 50,
          "type": "string"
        },
        "lastName": {
          "format": "string",
          "maxLength": 50,
          "type": "string"
        },
        "address": {
          "format": "string",
          "maxLength": 150,
          "type": "string"
        },
        "contactMobile": {
          "format": "tel",
          "maxLength": 10,
          "minLength": 9,
          "type": "string"
        },
        "imageProfile": {
          "type": "string"
        },
        "userTypeId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Bearer 12345abcdef"
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}